Localize host_check.check_clang in libcodechecker.analyze
