+++ libcodechecker/analyze/host_check.py
@@ -1,0 +1,37 @@
+# -------------------------------------------------------------------------
+#                     The CodeChecker Infrastructure
+#   This file is distributed under the University of Illinois Open Source
+#   License. See LICENSE.TXT for details.
+# -------------------------------------------------------------------------
+
+import errno
+import subprocess
+
+from libcodechecker.logger import LoggerFactory
+
+LOG = LoggerFactory.get_new_logger('HOST CHECK')
+
+
+def check_clang(compiler_bin, env):
+    """
+    Simple check if clang is available.
+    """
+    clang_version_cmd = [compiler_bin, '--version']
+    LOG.debug_analyzer(' '.join(clang_version_cmd))
+    try:
+        res = subprocess.call(clang_version_cmd,
+                              env=env,
+                              stdout=subprocess.PIPE,
+                              stderr=subprocess.PIPE)
+        if not res:
+            return True
+
+        LOG.debug_analyzer('Failed to run: "' + ' '.join(clang_version_cmd) +
+                           '"')
+        return False
+
+    except OSError as oerr:
+        if oerr.errno == errno.ENOENT:
+            LOG.error(oerr)
+            LOG.error('Failed to run: "' + ' '.join(clang_version_cmd) + '"')
+            return False
